+++ CHANGELOG.md
@@ -59,7 +59,7 @@
 
 ## 1.9.0
 - add WasmJS target
-- add WasmWasi target (not or KmmResult-test, as Kotest does not support WASI yet)
+- add WasmWasi target (not for KmmResult-test, as Kotest does not support WASI yet)
 - Function Renames (old ones are still present, but deprecated)
 - rename `wrapping` -> `catchingAs` but keep the old names as deprecated alternative
   - add `catchingUnwrappedAs`, which works just like `catchingAs` but on a `Result` rather than a `KmmResult` to avoid instantiation overhead